# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for file processing system with batch and item management",
        "title": "File Processing System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@fileprocessing.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/files": {
            "get": {
                "description": "Get a list of all uploaded files",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "List all files",
                "responses": {
                    "200": {
                        "description": "files, count",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/files/{id}": {
            "get": {
                "description": "Get the processing status and details of a file",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Get file status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FileMetadata"
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/files/{id}/download": {
            "get": {
                "description": "Download the generated PDF file for a completed file",
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Download PDF file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "File not found or PDF not available",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/scheduled": {
            "get": {
                "description": "Get a list of files that are scheduled for processing",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "List scheduled files",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Limit number of results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "files, count",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "description": "Upload a CSV or XLSX file for processing. Can be processed immediately or scheduled for later.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Upload a file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "immediate",
                            "scheduled"
                        ],
                        "type": "string",
                        "default": "immediate",
                        "description": "Schedule type: immediate or scheduled",
                        "name": "schedule_type",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Scheduled time in RFC3339 format (e.g., 2025-12-05T15:00:00Z). Required if schedule_type=scheduled, must be in the future",
                        "name": "scheduled_at",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "id, file_name, status, schedule_type, scheduled_at",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "File with this name already exists",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.FileMetadata": {
            "type": "object",
            "properties": {
                "bucket_name": {
                    "type": "string"
                },
                "content_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "object_name": {
                    "type": "string"
                },
                "pdf_path": {
                    "type": "string"
                },
                "processed_at": {
                    "type": "string"
                },
                "schedule_type": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "status": {
                    "description": "Changed to status_code",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}